+++ packages/cheqd-blockchain-modules/tests/blob-module.test.js
@@ -6,7 +6,7 @@ import {
 import { MultiApiBlobModule } from "@docknetwork/credential-sdk/modules";
 import generateBlobModuleTests from "@docknetwork/credential-sdk/modules/tests/blob-module";
 import CheqdDIDModule from "../src/did/module";
-import { faucet } from "./constants";
+import { faucet, url } from "./constants";
 import CheqdBlobModule from "../src/blob/module";
 import { MultiApiDIDModule } from "@docknetwork/credential-sdk/modules";
 
@@ -15,7 +15,7 @@ describe("BlobModule", () => {
 
   beforeAll(async () => {
     await cheqd.init({
-      url: process.env.ENDPOINT_URL || "http://localhost:26657",
+      url,
       mnemonic: faucet.mnemonic,
       network: "testnet",
     });