+++ packages/cheqd-blockchain-modules/tests/constants.js
@@ -1,7 +1,9 @@
 export const faucet = {
   prefix: "cheqd",
   minimalDenom: "ncheq",
-  mnemonic:
+  mnemonic: process.env.CHEQD_MNEMONIC ||
     "steak come surprise obvious remain black trouble measure design volume retreat float coach amused match album moment radio stuff crack orphan ranch dose endorse",
-  address: "cheqd1fgl67nvjdkrnaemjzg5sqvck9fcst4vt99gmma",
+  address: process.env.CHEQD_SENDER_ADDRESS || "cheqd1fgl67nvjdkrnaemjzg5sqvck9fcst4vt99gmma",
 };
+
+export const url = process.env.CHEQD_RPC_URL || "http://localhost:26657";